FunctionApp/requirements.txt
pandas
numpy
scikit-learn
azure-functions
azure-storage-blob